requirements.txt
numpy
matplotlib
pillow
tensorflow
streamlit
streamlit_extras
